refactor: move MultiSelect to own folder
